#96: (WIP) Tutorial updated with goomba for collision showcase.

+++ lionengine-examples/src/main/java/com/b3dgs/lionengine/example/game/state/StateTurn.java
@@ -39,7 +39,7 @@ class StateTurn
     /** Movement force. */
     private final Force movement;
     /** Movement side. */
-    private int side;
+    private double side;
 
     /**
      * Create the walk state.
